requirements.txt
django
gunicorn
whitenoise
psycopg[binary]
requests
urllib3
dj_database_url
django-filter
python-decouple
django-cors-headers
djangorestframework
pyspellchecker
pymupdf
google-generativeai
dotenv
supabase
django-prometheus
numpy>=1.23
pandas>=2.0
scikit-learn==1.6.0
joblib==1.3.2
nemoguardrails